# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Go Auth API",
    "description": "API аутентификации и авторизации пользователей с поддержкой OAuth2",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API server"
    }
  ],
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Обновление access токена",
        "description": "Получение нового access токена по refresh токену",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Токен успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Невалидный или истекший refresh токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/oauth/{provider}": {
      "get": {
        "tags": ["OAuth2"],
        "summary": "ШАГ 2-3: Получить URL авторизации провайдера",
        "description": "**Что происходит:**\n1. Client запрашивает ссылку для авторизации через провайдера\n2. Auth Server генерирует URL с параметрами (client_id, redirect_uri, scope, state)\n3. Auth Server возвращает URL для перенаправления пользователя\n\n**State** - случайная строка для защиты от CSRF атак",
        "operationId": "oauthGetAuthUrl",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "Провайдер OAuth2",
            "schema": {
              "$ref": "#/components/schemas/OAuthProvider"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": false,
            "description": "URL для возврата после авторизации (по умолчанию из конфига)",
            "schema": {
              "type": "string",
              "format": "uri",
              "example": "http://localhost:5173/auth/callback"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "URL для перенаправления пользователя на страницу провайдера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAuthUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неподдерживаемый провайдер",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oauth/{provider}/callback": {
      "get": {
        "tags": ["OAuth2"],
        "summary": "ШАГ 9-13: Callback после авторизации у провайдера",
        "description": "**Что происходит:**\n1. Пользователь подтвердил доступ у провайдера (Google/GitHub/etc)\n2. Провайдер редиректит сюда с authorization code\n3. Auth Server обменивает code на access_token провайдера\n4. Auth Server получает данные пользователя (email, name) от провайдера\n5. Auth Server создаёт/находит пользователя в БД\n6. Auth Server генерирует свои JWT токены\n7. Возвращает токены клиенту\n\n**Важно:** Этот endpoint вызывается автоматически провайдером!",
        "operationId": "oauthCallback",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "Провайдер OAuth2",
            "schema": {
              "$ref": "#/components/schemas/OAuthProvider"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "description": "Authorization code от провайдера (одноразовый, живёт ~10 минут)",
            "schema": {
              "type": "string",
              "example": "4/0AX4XfWh..."
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "description": "State для защиты от CSRF (должен совпадать с отправленным)",
            "schema": {
              "type": "string",
              "example": "abc123xyz"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "description": "Код ошибки, если пользователь отклонил доступ",
            "schema": {
              "type": "string",
              "example": "access_denied"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Редирект на фронтенд с токенами в URL или cookie",
            "headers": {
              "Location": {
                "description": "URL фронтенда с параметрами",
                "schema": {
                  "type": "string",
                  "example": "http://localhost:5173/auth/callback?access_token=...&refresh_token=..."
                }
              },
              "Set-Cookie": {
                "description": "JWT токены в httpOnly cookies (опционально)",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Невалидный code или state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь отклонил доступ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "description": "Текст ошибки"
          }
        }
      },
...
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh токен",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "RefreshTokenResponse": {
        "type": "object",
        "required": ["access_token"],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Новый JWT access токен",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
...
      "OAuthProvider": {
        "type": "string",
        "enum": ["google", "github", "yandex", "vk"],
        "description": "Поддерживаемые OAuth2 провайдеры",
        "example": "google"
      },
      "OAuthAuthUrlResponse": {
        "type": "object",
        "required": ["auth_url", "state"],
        "properties": {
          "auth_url": {
            "type": "string",
            "format": "uri",
            "description": "URL для перенаправления пользователя на страницу провайдера",
            "example": "https://accounts.google.com/o/oauth2/v2/auth?client_id=...&redirect_uri=...&scope=email%20profile&state=abc123&response_type=code"
          },
          "state": {
            "type": "string",
            "description": "Случайный state для CSRF защиты (сохраните и проверьте в callback)",
            "example": "abc123xyz789"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен в заголовке Authorization"
      },
      "OAuth2Google": {
        "type": "oauth2",
        "description": "OAuth2 авторизация через Google",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
              "email": "Доступ к email пользователя",
              "profile": "Доступ к профилю пользователя"
            }
          }
        }
      },
      "OAuth2GitHub": {
        "type": "oauth2",
        "description": "OAuth2 авторизация через GitHub",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://github.com/login/oauth/authorize",
            "tokenUrl": "https://github.com/login/oauth/access_token",
            "scopes": {
              "user:email": "Доступ к email пользователя",
              "read:user": "Доступ к профилю пользователя"
            }
          }
        }
      }
    }
  },
...
}